Project_manager_py/cursor_main_1/Python_Backend/requirements.txt
fastapi
uvicorn
asyncpg
python-dotenv 